*Version 2.0.12* * Cleaned up errors for accessing reports that no longer exist * Changed report instance delete logic to only delete emailed reports that are run off of a schedule * Added notes to help about new delete logic

+++ Reports/src/com/serotonin/m2m2/reports/web/ReportWorkItem.java
@@ -218,9 +218,11 @@ public class ReportWorkItem implements WorkItem {
             catch (AddressException e) {
                 LOG.error(e);
             }
-
-            // Delete the report instance.
-            reportDao.deleteReportInstance(reportInstance.getId(), user.getId());
+            
+            if(reportConfig.isSchedule()){
+	            // Delete the report instance.
+	            reportDao.deleteReportInstance(reportInstance.getId(), user.getId());
+            }
         }
 
         LOG.debug("Finished running report with id " + reportConfig.getId() + ", instance id " + reportInstance.getId());
